requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pandas==2.1.3
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Machine Learning
scikit-learn==1.3.2
prophet==1.1.5
tensorflow==2.15.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
black==23.11.0
isort==5.12.0 